Create a new hook for validating a signature
This hook effectively just wrap's viem's `publicClient`'s `verifyMessage` call, which takes a message, signature, and address, and (asynchronously) returns a boolean indicating if the data aligns. It's async because this `verifyMessage` function supports EIP1271, which needs to make an onchain request when the given address is a smart contract.