Use the new hook for signing Solo message
In `SoloNodeRequestAccess`, we'll no longer verify signatures by recovering the address from the signature and message, and comparing that recovered address to the connected wallet address. Instead, we'll just pass the message, signature, and connected address to the new hook and receive the boolean result.

This commit also makes the necessary changes to our "step" determination logic (truthy response instead of recovered address === connected address), and also simply passes the connected address (instead of the recovered address, which we don't have any more) into the `SignedMessageForm`.