Continue with new signature validation pattern
There's another hook called `useIsValidSignedMessage` which also performs signature validation. This commit replaces the old method of signature validation (by recovering an address from signature and message, and comparing it to the given address), with the new `verifyMessage` function which supports EIP1271 smart contract signatures.

I'm not quite sure why this separate hook needs to exists (it seems a little redundant?), but that's probably just beacuse of my lack of context.