Show "verifying" state in `SignedMessageForm`
This isn't strictly necessary for the work in this PR, but since the the new `verifyMessage` function is async, there's a moment after the `SignedMessageForm` appears but before signature validation in that component is complete, which was resulting in the "Error" message on the readonly input to appear. This commit just adds a third "verifying" state, so instead of a brief "error" message appearing while validation is ongoing, we see a "validating" message instead.