Keep message timestamp fresh
Create a new message (with the current timestamp) every second, until the user actually is ready to "Sign". This will ensure that they're not signing with a stale timestamp which is rejected from the API.

There is still a potential issue of the user not accepting the T&C and finally clicking the "Request Access" button in a reasonable amount of time.